I-#415-renamed dto and interfaces
